refactor: split storage modules
